+++ lib/nordic_nrf_mesh.dart
@@ -5,4 +5,4 @@ export 'src/ble/ble_manager.dart';
 export 'src/ble/ble_manager_callbacks.dart';
 export 'src/ble/ble_mesh_manager.dart';
 export 'src/ble/ble_mesh_manager_callbacks.dart';
-export 'src/utils/ble_mesh_manager_helper.dart';
+export 'src/utils/provisioning.dart';
